# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SN Auto Parts API
  description: |
    Feature-config driven e-commerce API for auto parts.
    
    ## Authentication
    All authenticated endpoints require a session token passed via:
    - Cookie: `auth-token=<token>`
    - Header: `Authorization: Bearer <token>`
    
    ## Authorization
    APIs are generic - authorization is determined by the authenticated user's role 
    and feature configuration. The same endpoint may return different data or 
    permissions based on the user's role.
    
    ## Feature-Based Access
    Each endpoint requires specific features to be enabled in the user's role configuration.
    If a user lacks the required feature, they receive a 403 Forbidden response.
  version: 1.0.0
  contact:
    name: SN Auto Parts
    email: support@snautoparts.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.snautoparts.com/api/v1
    description: Production server
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      operationId: listUsers
      security:
        - sessionAuth: []
      x-feature-required: users.viewAll
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
    post:
      tags: [Users]
      summary: Create user
      operationId: createUser
      security:
        - sessionAuth: []
      x-feature-required: users.create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/User'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user
      operationId: getUser
      security:
        - sessionAuth: []
      x-feature-required: users.viewAll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Users]
      summary: Update user role
      operationId: updateUserRole
      security:
        - sessionAuth: []
      x-feature-required: users.updateRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [CUSTOMER, MANAGER, ADMIN]
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      tags: [Users]
      summary: Delete user
      operationId: deleteUser
      security:
        - sessionAuth: []
      x-feature-required: users.delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ============================================
  # PROFILE ENDPOINTS
  # ============================================
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session token (also accepted as auth-token cookie)
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    Success:
      description: Operation successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    DataWrapper:
      type: object
      properties:
        data: {}
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
        details:
          type: object

    # Auth Schemas
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [CUSTOMER, MANAGER, ADMIN]
        image:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [CUSTOMER, MANAGER, ADMIN]
...